# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shop Backend API
  description: API completa para gestión de usuarios, productos, pedidos, carrito, pagos y notificaciones
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar un nuevo usuario
      description: Registra un usuario en la base de datos. No requiere autenticación.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Usuario registrado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario registrado correctamente"
        '400':
          description: Error de validación o negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión (login)
      description: Autentica usuario y devuelve un JWT y datos públicos del usuario.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login correcto — token y datos de usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: Credenciales inválidas u otro error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # -------------------- USERS --------------------
...
  /addresses:
    get:
      summary: Obtener todas las direcciones del usuario autenticado
      description: Devuelve una lista con todas las direcciones asociadas al usuario autenticado mediante el token JWT.
      tags: [Addresses]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de direcciones del usuario.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        '401':
          description: Token inválido o no proporcionado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Crear una nueva dirección para el usuario autenticado
      description: Crea una dirección asociada al usuario autenticado.
      tags: [Addresses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreateRequest'
      responses:
        '201':
          description: Dirección creada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressCreateResponse'
        '400':
          description: Error de validación o campos faltantes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /addresses/{id}:
    put:
      summary: Actualizar una dirección del usuario autenticado
      description: Permite modificar los datos de una dirección específica del usuario autenticado.
      tags: [Addresses]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressUpdateRequest'
      responses:
        '200':
          description: Dirección actualizada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressUpdateResponse'
        '400':
          description: Error de validación o dirección no encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Eliminar una dirección del usuario autenticado
      description: Elimina una dirección específica asociada al usuario autenticado.
      tags: [Addresses]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dirección eliminada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDeleteResponse'
        '400':
          description: Error al eliminar la dirección.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # -------------------- PRODUCTS --------------------
...
  /cart:
    get:
      summary: Obtener el carrito del usuario autenticado
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Carrito obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          description: No autenticado

    delete:
      summary: Vaciar completamente el carrito del usuario
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Carrito vaciado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Carrito eliminado correctamente"
        '401':
          description: No autenticado
  /cart/items:
    post:
      summary: Añadir un producto al carrito
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, quantity]
              properties:
                product_id:
                  type: integer
                  example: 15
                quantity:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Item añadido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Error al añadir item (por ejemplo stock insuficiente)
  /cart/items/{id}:
    put:
      summary: Actualizar cantidad de un item del carrito
      tags: [Cart]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del item del carrito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
                  example: 3
      responses:
        '200':
          description: Cantidad actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Error al actualizar cantidad

    delete:
      summary: Eliminar un item del carrito
      tags: [Cart]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del item del carrito
      responses:
        '200':
          description: Item eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Error al eliminar item
  /cart/checkout:
    post:
      summary: Realizar checkout del carrito (crea orden + sesión de pago)
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_id, frontendUrl]
              properties:
                address_id:
                  type: integer
                  example: 4
                frontendUrl:
                  type: string
                  example: "https://miapp.com"
      responses:
        '200':
          description: Checkout iniciado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: integer
                    example: 123
                  url:
                    type: string
                    example: "https://checkout.stripe.com/session/abc123"
        '400':
          description: Error al iniciar checkout

  # -------------------- ORDERS --------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AuthRegisterRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - password
      properties:
        first_name:
          type: string
          example: "María"
        last_name:
          type: string
          example: "Pérez"
        email:
          type: string
          format: email
          example: "maria@example.com"
        password:
          type: string
          example: "contraseñaSegura123"
        role:
          type: string
          description: "Opcional: 'user' o 'admin' (si lo permites)."
          example: "user"
        phone:
          type: string
          example: "+34 600 000 000"
    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "maria@example.com"
        password:
          type: string
          example: "contraseñaSegura123"
    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT para autenticación (bearer)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/UserPublic'
    UserPublic:
      type: object
      properties:
        id:
          type: integer
          example: 12
        first_name:
          type: string
          example: "María"
        last_name:
          type: string
          example: "Pérez"
        email:
          type: string
          format: email
          example: "maria@example.com"
        role:
          type: string
          example: "user"
        phone:
          type: string
          example: "+34 600 000 000"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Error message"
        error:
          type: string
          nullable: true

    # -------------------- USERS --------------------
...
    Address:
      type: object
      properties:
        id:
          type: integer
          example: 3
        first_name:
          type: string
          example: "María"
        last_name:
          type: string
          example: "Pérez"
        street:
          type: string
          example: "Calle Mayor 12, 3ºA"
        city:
          type: string
          example: "Madrid"
        province:
          type: string
          example: "Madrid"
        postal_code:
          type: string
          example: "28013"
        country:
          type: string
          example: "España"
        phone:
          type: string
          example: "+34 600 123 456"
        type:
          type: string
          example: "shipping"
        created_at:
          type: string
          format: date-time
          example: "2025-03-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-03-01T12:00:00Z"
    AddressCreateRequest:
      type: object
      required:
        - first_name
        - last_name
        - street
        - city
        - province
        - postal_code
        - country
      properties:
        first_name:
          type: string
          example: "María"
        last_name:
          type: string
          example: "Pérez"
        street:
          type: string
          example: "Calle Mayor 12, 3ºA"
        city:
          type: string
          example: "Madrid"
        province:
          type: string
          example: "Madrid"
        postal_code:
          type: string
          example: "28013"
        country:
          type: string
          example: "España"
        phone:
          type: string
          example: "+34 600 123 456"
        type:
          type: string
          example: "billing"
    AddressCreateResponse:
      type: object
      properties:
        message:
          type: string
          example: "Dirección añadida correctamente"
    AddressUpdateRequest:
      type: object
      properties:
        first_name:
          type: string
          example: "Ana"
        last_name:
          type: string
          example: "Gómez"
        street:
          type: string
          example: "Avenida del Sol 45"
        city:
          type: string
          example: "Barcelona"
        province:
          type: string
          example: "Cataluña"
        postal_code:
          type: string
          example: "08025"
        country:
          type: string
          example: "España"
        phone:
          type: string
          example: "+34 611 222 333"
        type:
          type: string
          example: "shipping"
    AddressUpdateResponse:
      type: object
      properties:
        message:
          type: string
          example: "Dirección actualizada correctamente"
    AddressDeleteResponse:
      type: object
      properties:
        message:
          type: string
          example: "Dirección eliminada correctamente"

    # -------------------- PRODUCTS --------------------
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
          example: 12
        product_id:
          type: integer
          example: 5
        product_name:
          type: string
          example: "Camiseta básica"
        price:
          type: number
          format: float
          example: 19.99
        quantity:
          type: integer
          example: 2
        stock:
          type: integer
          nullable: true
          example: 10
        image_url:
          type: string
          nullable: true
          example: "https://api.miapp.com/uploads/123-camiseta.jpg"
        subtotal:
          type: number
          format: float
          example: 39.98
    CartResponse:
      type: object
      properties:
        cart_id:
          type: integer
          example: 7
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        total:
          type: number
          format: float
          example: 59.97

    # -------------------- ORDERS --------------------
...